+++ Core/include/Acts/Utilities/Concepts.hpp
@@ -22,37 +22,5 @@
 #define ACTS_CONCEPT(x) typename
 #define ACTS_STATIC_CHECK_CONCEPT(concept, type) \
   static_assert(true, "Dummy assertion")
-#endif
-
-#if defined(__cpp_concepts)
-// Define some concepts
-
-#include <ranges>
-
-namespace Acts {
-
-template <typename external_t>
-concept isCollectionThatSupportsPushBack =
-    std::ranges::range<external_t> && requires {
-  typename external_t::value_type;
-} && requires(external_t coll, typename external_t::value_type val) {
-  coll.push_back(val);
-};
-
-template <typename external_t>
-concept isCollectionThatSupportsInsert =
-    std::ranges::range<external_t> && requires {
-  typename external_t::value_type;
-} && requires(external_t coll, typename external_t::value_type val) {
-  coll.insert(std::ranges::end(coll), val);
-};
-
-template <typename external_t, typename... args_t>
-concept isCollectionThatSupportsEmplace = std::ranges::range<external_t> &&
-    requires(external_t coll, args_t... vals) {
-  coll.emplace_back(vals...);
-};
-
-}  // namespace Acts
 
 #endif
