+++ Tests/UnitTests/Core/Seeding/CMakeLists.txt
@@ -4,3 +4,4 @@ target_link_libraries(ActsUnitTestSeedFinder PRIVATE ActsCore Boost::boost)
 add_unittest(EstimateTrackParamsFromSeed EstimateTrackParamsFromSeedTest.cpp)
 add_unittest(BinnedGroupTest BinnedGroupTest.cpp)
 add_unittest(HoughTransformTest HoughTransformTest.cpp)
+add_unittest(UtilityFunctions UtilityFunctionsTests.cpp)
