+++ Tests/UnitTests/Core/Seeding/UtilityFunctionsTests.cpp
@@ -8,7 +8,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "Acts/Utilities/Functions.hpp"
+#include "Acts/Seeding/detail/Functions.hpp"
 
 #include <iterator>
 #include <list>
@@ -22,12 +22,12 @@ namespace Acts::Test {
 #if defined(__cpp_concepts)
 BOOST_AUTO_TEST_CASE(insert_vector) {
   std::vector<std::size_t> coll;
-  Acts::Utils::insert(coll, 2ul);
+  Acts::details::insert(coll, 2ul);
   BOOST_CHECK(coll.size() == 1ul);
-  Acts::Utils::insert(coll, 5ul);
+  Acts::details::insert(coll, 5ul);
   BOOST_CHECK(coll.size() == 2ul);
   std::size_t val = 1ul;
-  Acts::Utils::insert(coll, val);
+  Acts::details::insert(coll, val);
   BOOST_CHECK(coll.size() == 3ul);
 
   BOOST_CHECK_EQUAL(coll[0], 2ul);
@@ -37,12 +37,12 @@ BOOST_AUTO_TEST_CASE(insert_vector) {
 
 BOOST_AUTO_TEST_CASE(insert_list) {
   std::list<std::size_t> coll;
-  Acts::Utils::insert(coll, 2ul);
+  Acts::details::insert(coll, 2ul);
   BOOST_CHECK(coll.size() == 1ul);
-  Acts::Utils::insert(coll, 5ul);
+  Acts::details::insert(coll, 5ul);
   BOOST_CHECK(coll.size() == 2ul);
   std::size_t val = 1ul;
-  Acts::Utils::insert(coll, val);
+  Acts::details::insert(coll, val);
   BOOST_CHECK(coll.size() == 3ul);
 
   BOOST_CHECK_EQUAL(coll.front(), 2ul);
@@ -55,12 +55,12 @@ BOOST_AUTO_TEST_CASE(insert_list) {
 
 BOOST_AUTO_TEST_CASE(insert_set) {
   std::set<std::size_t> coll;
-  Acts::Utils::insert(coll, 2ul);
+  Acts::details::insert(coll, 2ul);
   BOOST_CHECK(coll.size() == 1ul);
-  Acts::Utils::insert(coll, 5ul);
+  Acts::details::insert(coll, 5ul);
   BOOST_CHECK(coll.size() == 2ul);
   std::size_t val = 1ul;
-  Acts::Utils::insert(coll, val);
+  Acts::details::insert(coll, val);
   BOOST_CHECK(coll.size() == 3ul);
 
   BOOST_CHECK(coll.find(2ul) != coll.end());
@@ -70,12 +70,12 @@ BOOST_AUTO_TEST_CASE(insert_set) {
 
 BOOST_AUTO_TEST_CASE(insert_unordered_set) {
   std::unordered_set<std::size_t> coll;
-  Acts::Utils::insert(coll, 2ul);
+  Acts::details::insert(coll, 2ul);
   BOOST_CHECK(coll.size() == 1ul);
-  Acts::Utils::insert(coll, 5ul);
+  Acts::details::insert(coll, 5ul);
   BOOST_CHECK(coll.size() == 2ul);
   std::size_t val = 1ul;
-  Acts::Utils::insert(coll, val);
+  Acts::details::insert(coll, val);
   BOOST_CHECK(coll.size() == 3ul);
 
   BOOST_CHECK(coll.find(2ul) != coll.end());
@@ -86,12 +86,12 @@ BOOST_AUTO_TEST_CASE(insert_unordered_set) {
 BOOST_AUTO_TEST_CASE(insert_back_iterator) {
   std::vector<std::size_t> coll;
   std::back_insert_iterator<decltype(coll)> backItr(coll);
-  Acts::Utils::insert(backItr, 2ul);
+  Acts::details::insert(backItr, 2ul);
   BOOST_CHECK(coll.size() == 1ul);
-  Acts::Utils::insert(backItr, 5ul);
+  Acts::details::insert(backItr, 5ul);
   BOOST_CHECK(coll.size() == 2ul);
   std::size_t val = 1ul;
-  Acts::Utils::insert(backItr, val);
+  Acts::details::insert(backItr, val);
   BOOST_CHECK(coll.size() == 3ul);
 
   BOOST_CHECK_EQUAL(coll[0], 2ul);
@@ -109,12 +109,12 @@ BOOST_AUTO_TEST_CASE(emplace_vector) {
   };
 
   std::vector<A> coll;
-  Acts::Utils::emplace(coll, 2ul, 4.5f, "A");
+  Acts::details::emplace(coll, 2ul, 4.5f, "A");
   BOOST_CHECK(coll.size() == 1ul);
   std::size_t a = 5ul;
   float b = 0.25f;
   std::string c = "B";
-  Acts::Utils::emplace(coll, a, b, c);
+  Acts::details::emplace(coll, a, b, c);
   BOOST_CHECK(coll.size() == 2ul);
 
   BOOST_CHECK_EQUAL(coll[0].valA, 2ul);
@@ -129,12 +129,12 @@ BOOST_AUTO_TEST_CASE(emplace_vector) {
 BOOST_AUTO_TEST_CASE(insert_back_iterator) {
   std::vector<std::size_t> coll;
   std::back_insert_iterator<decltype(coll)> backItr(coll);
-  Acts::Utils::insert(backItr, 2ul);
+  Acts::details::insert(backItr, 2ul);
   BOOST_CHECK(coll.size() == 1ul);
-  Acts::Utils::insert(backItr, 5ul);
+  Acts::details::insert(backItr, 5ul);
   BOOST_CHECK(coll.size() == 2ul);
   std::size_t val = 1ul;
-  Acts::Utils::insert(backItr, val);
+  Acts::details::insert(backItr, val);
   BOOST_CHECK(coll.size() == 3ul);
 
   BOOST_CHECK_EQUAL(coll[0], 2ul);
