+++ Tests/UnitTests/Core/Utilities/CMakeLists.txt
@@ -14,7 +14,6 @@ add_unittest(DBScan DBScanTests.cpp)
 add_unittest(Extendable ExtendableTests.cpp)
 add_unittest(FiniteStateMachine FiniteStateMachineTests.cpp)
 add_unittest(Frustum FrustumTest.cpp)
-add_unittest(Functions FunctionsTests.cpp)
 add_unittest(Grid GridTests.cpp)
 add_unittest(GridAccessHelpers GridAccessHelpersTests.cpp)
 add_unittest(GridAxisGenerators GridAxisGeneratorsTests.cpp)
